+++ working_flask/templates/index.html
@@ -5,17 +5,17 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Basketball!</title>
+  <title>Superheroes!</title>
   <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css">
 </head>
 
 <body>
   <div class="container text-center">
-    <h1 class="jumbotron">Basketball Players</h1>
+    <h1 class="jumbotron">Superheroes</h1>
     <div>
       <ul style="list-style: none;">
-        {% for team in teams %}
-          <li>{{ team.player }}: {{ team.position }}</li>
+        {% for superhero in superheroes %}
+          <li>{{ superhero.hero }}: {{ superhero.power }}</li>
         {% endfor %}
       </ul>
     </div>
